chore(6932): convert final legacy context consumers to useI18nContext for React 19
