+++ ui/components/app/error-boundary/error-boundary.tsx
@@ -1,9 +1,14 @@
 import React, { Component, ReactNode, ErrorInfo } from 'react';
 import { captureException } from '../../../../shared/lib/sentry';
-import { I18nContext } from '../../../contexts/i18n';
+import { useI18nContext } from '../../../hooks/useI18nContext';
 
 type ErrorBoundaryProps = { children: ReactNode };
 type ErrorBoundaryState = { hasError: boolean };
+
+export function ErrorFallback() {
+  const t = useI18nContext();
+  return <p className="p-4 text-center">{t('somethingWentWrong')}</p>;
+}
 
 class ErrorBoundary extends Component<ErrorBoundaryProps, ErrorBoundaryState> {
   constructor(props: ErrorBoundaryProps) {
@@ -21,11 +26,7 @@ class ErrorBoundary extends Component<ErrorBoundaryProps, ErrorBoundaryState> {
 
   render() {
     if (this.state.hasError) {
-      return (
-        <I18nContext.Consumer>
-          {(t) => <p className="p-4 text-center">{t('somethingWentWrong')}</p>}
-        </I18nContext.Consumer>
-      );
+      return <ErrorFallback />;
     }
 
     return this.props.children;
