--- ui/components/app/modals/hide-token-confirmation-modal/hide-token-confirmation-modal.js
@@ -1,222 +1,0 @@
-import PropTypes from 'prop-types';
-import React, { Component } from 'react';
-import { connect } from 'react-redux';
-import {
-  formatChainIdToCaip,
-  isNonEvmChainId,
-} from '@metamask/bridge-controller';
-import {
-  AvatarToken,
-  AvatarTokenSize,
-  Box,
-  BoxAlignItems,
-  BoxJustifyContent,
-} from '@metamask/design-system-react';
-import { I18nContext } from '../../../../contexts/i18n';
-import * as actions from '../../../../store/actions';
-import { Button, ButtonVariant } from '../../../component-library';
-import { DEFAULT_ROUTE } from '../../../../helpers/constants/routes';
-import {
-  getCurrentChainId,
-  getNetworkConfigurationsByChainId,
-} from '../../../../../shared/lib/selectors/networks';
-import { getInternalAccountBySelectedAccountGroupAndCaip } from '../../../../selectors/multichain-accounts/account-tree';
-import { toAssetId } from '../../../../../shared/lib/asset-utils';
-import { getIsAssetsUnifyStateEnabled } from '../../../../selectors/assets-unify-state/feature-flags';
-import {
-  getAssetsControllerCustomAssets,
-  isAssetInAccountCustomAssets,
-} from '../../../../selectors/assets-unify-state/asset-preferences';
-
-function mapStateToProps(state) {
-  return {
-    chainId: getCurrentChainId(state),
-    token: state.appState.modal.modalState.props.token,
-    navigate: state.appState.modal.modalState.props.navigate,
-    networkConfigurationsByChainId: getNetworkConfigurationsByChainId(state),
-    getAccountForChain: (caipChainId) =>
-      getInternalAccountBySelectedAccountGroupAndCaip(state, caipChainId),
-    assetsUnifyStateFeatureEnabled: getIsAssetsUnifyStateEnabled(state),
-    customAssets: getAssetsControllerCustomAssets(state),
-  };
-}
-
-function mapDispatchToProps(dispatch) {
-  return {
-    hideModal: () => dispatch(actions.hideModal()),
-    hideToken: async (
-      address,
-      networkClientId,
-      chainId,
-      getAccountForChain,
-      assetsUnifyStateFeatureEnabled,
-      customAssets,
-    ) => {
-      const isNonEvm = isNonEvmChainId(chainId);
-
-      if (assetsUnifyStateFeatureEnabled) {
-        const assetId = toAssetId(address, chainId);
-        const caipChainId = isNonEvmChainId(chainId)
-          ? chainId
-          : formatChainIdToCaip(chainId);
-        const accountForChain = getAccountForChain(caipChainId);
-        const isInCustomAssets =
-          accountForChain &&
-          isAssetInAccountCustomAssets(
-            customAssets,
-            accountForChain.id,
-            assetId,
-          );
-
-        try {
-          if (isInCustomAssets) {
-            await dispatch(
-              actions.removeCustomAsset(accountForChain.id, assetId),
-            );
-          } else {
-            await dispatch(actions.hideAsset(assetId));
-          }
-        } catch (error) {
-          console.error('Error hiding/removing asset:', error);
-          return;
-        }
-      }
-
-      if (isNonEvm) {
-        // Handle non-EVM tokens
-        const accountForChain = getAccountForChain(chainId);
-
-        if (!accountForChain) {
-          console.warn(`No account found for chain ${chainId}`);
-          return;
-        }
-
-        await dispatch(
-          actions.multichainIgnoreAssets([address], accountForChain.id),
-        );
-      } else {
-        await dispatch(
-          actions.ignoreTokens({
-            tokensToIgnore: address,
-            networkClientId,
-            chainId,
-          }),
-        );
-      }
-
-      dispatch(actions.hideModal());
-    },
-  };
-}
-
-class HideTokenConfirmationModal extends Component {
-  static contextType = I18nContext;
-
-  static propTypes = {
-    hideToken: PropTypes.func.isRequired,
-    hideModal: PropTypes.func.isRequired,
-    chainId: PropTypes.string.isRequired,
-    networkConfigurationsByChainId: PropTypes.object.isRequired,
-    getAccountForChain: PropTypes.func.isRequired,
-    assetsUnifyStateFeatureEnabled: PropTypes.bool.isRequired,
-    customAssets: PropTypes.object,
-    token: PropTypes.shape({
-      symbol: PropTypes.string,
-      address: PropTypes.string,
-      image: PropTypes.string,
-      chainId: PropTypes.string,
-    }),
-    navigate: PropTypes.func,
-  };
-
-  state = {};
-
-  render() {
-    const {
-      chainId,
-      token,
-      hideToken,
-      hideModal,
-      navigate,
-      networkConfigurationsByChainId,
-      getAccountForChain,
-      assetsUnifyStateFeatureEnabled,
-      customAssets,
-    } = this.props;
-    const { symbol, address, image, chainId: tokenChainId } = token;
-    const chainIdToUse = tokenChainId || chainId;
-
-    return (
-      <div className="hide-token-confirmation__container">
-        <div className="hide-token-confirmation__title">
-          {this.context('hideTokenPrompt')}
-        </div>
-        <AvatarToken
-          className="hide-token-confirmation__identicon"
-          size={AvatarTokenSize.Xl}
-          name={symbol || address}
-          src={image}
-        />
-        <div className="hide-token-confirmation__symbol">{symbol}</div>
-        <div className="hide-token-confirmation__copy">
-          {this.context('readdToken')}
-        </div>
-        <Box
-          className="flex w-full"
-          justifyContent={BoxJustifyContent.Center}
-          alignItems={BoxAlignItems.Center}
-          gap={4}
-          marginTop={4}
-        >
-          <Button
-            variant={ButtonVariant.Secondary}
-            block
-            data-testid="hide-token-confirmation__cancel"
-            onClick={() => hideModal()}
-          >
-            {this.context('cancel')}
-          </Button>
-          <Button
-            variant={ButtonVariant.Primary}
-            block
-            data-testid="hide-token-confirmation__hide"
-            onClick={() => {
-              if (isNonEvmChainId(chainIdToUse)) {
-                hideToken(
-                  address,
-                  undefined,
-                  chainIdToUse,
-                  getAccountForChain,
-                  assetsUnifyStateFeatureEnabled,
-                  customAssets,
-                );
-              } else {
-                const chainConfig =
-                  networkConfigurationsByChainId[chainIdToUse];
-                const { defaultRpcEndpointIndex } = chainConfig;
-                const { networkClientId: networkInstanceId } =
-                  chainConfig.rpcEndpoints[defaultRpcEndpointIndex];
-                hideToken(
-                  address,
-                  networkInstanceId,
-                  chainIdToUse,
-                  getAccountForChain,
-                  assetsUnifyStateFeatureEnabled,
-                  customAssets,
-                );
-              }
-              navigate(DEFAULT_ROUTE);
-            }}
-          >
-            {this.context('hide')}
-          </Button>
-        </Box>
-      </div>
-    );
-  }
-}
-
-export default connect(
-  mapStateToProps,
-  mapDispatchToProps,
-)(HideTokenConfirmationModal);
