+++ ui/components/app/permissions-connect-footer/permissions-connect-footer.component.tsx
@@ -1,0 +1,39 @@
+import React from 'react';
+import {
+  Box,
+  BoxAlignItems,
+  BoxJustifyContent,
+} from '@metamask/design-system-react';
+import { useI18nContext } from '../../../hooks/useI18nContext';
+import ZENDESK_URLS from '../../../helpers/constants/zendesk-url';
+import { TextVariant } from '../../../helpers/constants/design-system';
+import { ButtonLink, ButtonLinkSize, Text } from '../../component-library';
+
+export default function PermissionsConnectFooter() {
+  const t = useI18nContext();
+
+  return (
+    <Box
+      className="flex"
+      alignItems={BoxAlignItems.Center}
+      justifyContent={BoxJustifyContent.Center}
+    >
+      <Text variant={TextVariant.bodyMd}>
+        {t('onlyConnectTrust', [
+          <ButtonLink
+            key="permission-connect-footer-learn-more-link"
+            size={ButtonLinkSize.Inherit}
+            target="_blank"
+            onClick={() => {
+              global.platform.openTab({
+                url: ZENDESK_URLS.USER_GUIDE_DAPPS,
+              });
+            }}
+          >
+            {t('learnMoreUpperCase')}
+          </ButtonLink>,
+        ])}
+      </Text>
+    </Box>
+  );
+}
